[wallet-ext] fix toasts being hidden behind overlay screens (#9322)
## Description 

I introduced a bug the other day where toast messages were being hidden
on overlay screens. This PR makes a generic `Portal` component and
portals both the overlay and toaster components inside the main page
layout container. To get the correct stacking order, we're just going to
use `9999` for the toaster and `999` for the overlay (it would be nice
to figure out a better way to manage these types of stacking contexts in
the future so these types of bugs don't happen as frequently). I did
consider pulling out `toaster-portal-container` to the `body` of the
document, but that came with its own positioning and layout issues.

<img width="460" alt="image"
src="https://user-images.githubusercontent.com/7453188/225321997-dea14413-ef63-45e8-a33a-2aea0436d756.png">

## Test Plan 
- Opened an overlay and then kicked off a toast message
- Opened the account menu and played around there
- Tested in both full-screen and pop-up mode

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
N/A