[simtests] Stress test committee changes (#9374)
As in title. The test starts with a committee size of 5, and a standby
set of validator configs, of initial size 6, and at each epoch, requests
to add 2 validators from the standby set, and requests to remove 2
validators from the current committee, until all validators in the
standby set have been exhausted.

Note that we cannot test recycling the same validators in and out
without more effort due to issues in fully shutting down all validator
components when a node is demoted from a validator to a fullnode.

Note also that, due to the above, the test takes a long time to run, as,
at each new epoch, we need to start new validators from genesis, and
give time for them to catch up to the current epoch and reconfig as
validators. This takes successively longer as we progress further from
genesis. Currently we wait `30 sec * epochId`, but it's possible we can
reduce this number somewhat.