crypto: pass in hashed message for sign and verify (#9561)
## Description 

Pass in hashed message for sign and verify API for user signatures.
Closes https://github.com/MystenLabs/sui/issues/9512
## Test Plan 

Existing unit tests and e2e test. 
---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes