[gas] Switch default gas table to use tier-based model (#9672)
This switches the default gas model for Sui to use a tier-based gas
model. The old bytecode gas model is still kept so we can continue to
compare against it, but it is behind a compile-time feature flag.

This also updates tests. The exact cost table that is chosen here will
be refined over the coming days and _is not_ the final cost table, but
that will be in a separate PR.