[explorer] - add checkpoints table  (#9636)
## Description 

- Adds a tab for checkpoints to the home page / transactions tables
- Currently duplicates a lot of the logic from the RecentTxCard
component. I went down the road of cleaning things up, but for the sake
of time, submitting in this state


Question: 
- Since the Transactions table now also has a checkpoints tab, we should
probably update the route. I left it as-is for now.
- Some options: 
   -  leave as is for now
   - `/transactions` and `/checkpoints` route to the same place
   - `/activity/:type` `/latest/:type` `/recent/:type` etc. 


## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes