[wallet-kit-site] Rework build to use path aliases (#9781)
## Description 

Small tweak to how the wallet kit site is built, so that it uses path
aliases and doesn't depend on artifacts existing. I dislike path aliases
and liked our other approach with using import conditions, but that
seems impossible to get working with Next for some reason, so this will
have to do.

## Test Plan 

Ran site locally on clean checkout.