[fix] flaky test in proposer (#9809)
## Description 

This PR is fixing a flaky test in Narwhal and more specifically when
testing our gRPC Proposer API . Basically the order of the
AuthorityFixtures wasn't aligned with the produced Committee
authorities. That made the `test_rounds_errors` test fail randomly as it
is based on parsing both the Committee and AuthorityFixtures.

Also refactored the test to iterate over the authorities using a common
approach via the `fixture.authorities()`

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes