Add StakeStatus::Unstaked for Stakes that were withdrawn (#9832)
## Description 

sui_getStakesByIds will return StakeStatus::Unstaked if the Stake was
withdrawn. More specifically, if a provided staked_sui_id returns a
Deleted object, then we query the information of the previous version of
that object when it existed and display that stake information with
status Unstaked.

## Test Plan 

Add test_unstaking in rpc_server_tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes