[explorer] Rebuild transactions for address and object (#9852)
## Description 

On object and address pages, the transaction table used a deprecated
method and required calling multiGetTransactions to populate. This
caused it to fail for big objects / transactions. Rewriting onto native
queryTransactions. For now, this has the limitation of not being able to
paginate, because we need to interleave two result sets. But once the
API is updated to support this filter we can pretty easily change that.

## Test Plan 

Ran locally.