[validator] Introduce AuthorityServer::submit_certificate (#9967)
This PR introduce `submit_certificate` method to authority, that allows
to persist certificate in the validator DB, without waiting for it to be
executed. This is useful to provide an early finality guarantee to the
user in some cases.

When 2f+1 validators returned success on `submit_certificate` RPC, it is
guaranteed that transaction will be executed before end of the epoch.

Differences between `handle_certificate` and `submit_certificate`:

(1) Unlike with `handle_certificate`, the result of execution is not
always known when `submit_certificate` returns.

(2) `submit_certificate` has lower latency then `handle_certificate`.
The difference is significant for shared object transactions, as
`submit_certificate` does not wait for transaction to be sequenced in
consensus.

It is important to note, that if you want to ensure that you transaction
will be executed, you *need* to wait for 2f+1 successful responses from
`submit_certificate`. Even if you are sure that validator you are
submitting transaction to is not byzantine, single validator response
does not provide guarantee that transaction will be sequnced by end of
the epoch.