throw error on temp txs (#10113)
## Description 

Full Node's `get_transaction_block` tries to handle temp txs by checking
if a given transaction digest can be mapped to a checkpoint.

This leads to unexpected behavior where a digest for a nonexistent
transaction with options=None will return a success response. This is
because `get_transaction_block` doesn't actually check if the
transaction exists in the first place. We now unconditionally make a
`get_executed_transaction` check to avoid this state.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes