[Narwhal] sleep in proposer until max parent time (#10148)
## Description 

The header creation time is set later than the added sleep, so it becomes
less likely that a Header with inaccurate timestamp get proposed. We
still need https://github.com/MystenLabs/sui/pull/10145 to enforce the
invariant.

## Test Plan 

n/a

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes